Compliance mixin; Use it in the VM and Host
